bot/requirements.txt
# Shellia AI Bot - Requirements
# Version: 2.0-Security

discord.py>=2.3.0
supabase>=2.0.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
stripe>=7.0.0

# ============ SECURITY DEPENDENCIES ============

# Chiffrement des secrets (secure_config.py)
cryptography>=41.0.0

# Rate limiting persistant avec Redis (optional)
redis>=5.0.0

# Vault integration (optional - si HashiCorp Vault utilisé)
# hvac>=2.0.0

# Additional security utilities
pyOpenSSL>=24.0.0
certifi>=2024.0.0
